ScoreManager/requirements.txt
appnope==0.1.0
backcall==0.1.0
bleach==3.0.2
certifi==2018.8.24
chardet==3.0.4
decorator==4.3.0
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.0.7
django-allauth==0.38.0
django-cors-headers==2.4.0
django-filter==2.0.0
django-import-export==1.1.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
entrypoints==0.2.3
et-xmlfile==1.0.1
idna==2.7
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jdcal==1.4
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Markdown==2.6.11
MarkupSafe==1.1.0
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.2
oauthlib==2.1.0
odfpy==1.3.6
openpyxl==2.5.9
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.5.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
Pygments==2.3.0
PyJWT==1.6.4
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.19.1
requests-oauthlib==1.0.0
Send2Trash==1.5.0
six==1.11.0
tablib==0.12.1
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
unicodecsv==0.14.1
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
xlrd==1.1.0
xlwt==1.3.0
